mv db funcs into /server
